automatisation mail/requirements.txt
imaplib
email
openai
python-dotenv
tkinter
sqlite3
re
datetime
threading